+++ cardano-slotting/CHANGELOG.md
@@ -1,9 +1,19 @@
 # Changelog for `cardano-slotting`
 
-## 0.1.2.0
+## 0.2.0.0
 
+* Add `binOpEpochNo` helper function to facilitate binary operations on
+  `EpochNo`.
+* Remove numeric instances (`Num`, `Integral`, `Real`) of `EpochNo` and
+  `EpochSize` for safety. 
+  They are still available for testing from the `testlib` as orphans.
 * New `Test.Cardano.Slotting.TreeDiff` module extracted from
   `cardano-ledger-binary`. It lives in a new public sublibrary `testlib`.
+
+### `testlib`
+
+* Add numeric instances (`Num`, `Integral`, `Real`) of `EpochNo` and
+  `EpochSize` as orphans.
 
 ## 0.1.1.1